Logout Added
Logout( as a button) is added but need to setup navigation.

It adds a button to logout but the is that it should rather go to the nutrition page instead.
Need to fix navigation and save the login data in firebase.

Adding the data to firebase to setup login should be the next step.

Thinking of putting all google authentication files in a different .js file to make it re-useable.