Accessibility  Page
Completed the UI for the Accessibility page.
Text size changes with corresponding buttons and colourblind switch can change colours of elements, this has not yet been set for colourblind people just proof of working.

To import the variables in other screens you'll need to add "import {switchColour, isLarge} from "./Accessibility";"

Then for fontSize and colours that need to be changed use the format, "isLarge ? 30 : 24," where the text size or colour noramlly would be.