Accessibility screen pull (#35)
* Accessibility  Page

Completed the UI for the Accessibility page.

Only works on Create Account and 2Factor Authentication page

Text size changes with corresponding buttons and colourblind switch can change colours of elements, this has not yet been set for colourblind people just proof of working.

To import the variables in other screens you'll need to add "import {switchColour, isLarge} from "./Accessibility";"

Then for fontSize and colours that need to be changed use the format, "isLarge ? 30 : 24," where the text size or colour noramlly would be.

* Updated Accessibility screen

Can now send colourblind and enlarged text variables to other screens. Explained in Trello ticket #106 how to do this

* Accessibility

* Accessibility fix

The buttons on the accessibility screen should now align and same goes for permissions. Authentication and create account continue buttons should now turn red when colourblind mode is on