Update create account
Added the use of useEffect hook to handle changes to the switchColour and isLarge props passed from the parent component.
Changed the back arrow icon in the imports to the correct one.
Added useRoute hook from @react-navigation/native to access the switchColour and isLarge props passed from the previous screen.
Removed unused styles in the StyleSheet.
Moved the styles constant inside the CreateAccount function for better encapsulation and readability.
Replaced hard-coded styles with dynamic ones that take into account the isLarge and switchColour props.
Removed the emailValidationText style as it is not used anymore.
Added switchColour to the backgroundColor of the button style, depending on its value.
Updated the onPress function of the Continue button to pass the switchColour and isLarge props to the next screen.